+++ room/src/main/java/com/team23/room/data/AppDatabase.kt
@@ -3,6 +3,8 @@ package com.team23.room.data
 import androidx.room.Database
 import androidx.room.RoomDatabase
 import androidx.room.TypeConverters
+import androidx.room.migration.Migration
+import androidx.sqlite.db.SupportSQLiteDatabase
 import com.team23.room.data.AppDatabase.Companion.BDD_VERSION
 import com.team23.room.data.converters.DateConverter
 import com.team23.room.data.daos.AchievementDao
@@ -16,12 +18,29 @@ import com.team23.room.data.entities.SettingEntity
 
 @Database(
     entities = [AchievementEntity::class, FactEntity::class, CategoryEntity::class, SettingEntity::class],
-    version = BDD_VERSION
+    version = BDD_VERSION,
 )
 @TypeConverters(DateConverter::class)
 abstract class AppDatabase : RoomDatabase() {
     companion object {
-        const val BDD_VERSION = 1
+        const val BDD_VERSION = 2
+
+        val migrationFrom1To2 = object: Migration(1, 2) {
+            override fun migrate(database: SupportSQLiteDatabase) {
+                val cursor = database.query("SELECT DISTINCT id_fonc FROM T_FACT WHERE isNew = 0")
+                val readFactId: MutableList<Long> = mutableListOf()
+                while (cursor.moveToNext()) {
+                    readFactId.add(cursor.getLong(0))
+                }
+                println("23facts all read facts: $readFactId")
+                database.delete("T_FACT", null, null)
+                database.execSQL(Scripts.factsScript1)
+                database.execSQL(Scripts.factsScript2)
+                readFactId.forEach { factId ->
+                    database.execSQL("UPDATE T_FACT SET isNew = 0 WHERE id_fonc = :factId")
+                }
+            }
+        }
     }
 
     abstract fun achievementsDao(): AchievementDao